Attempted a logging framework that failed. Uploading this state for version control before creating a new bbranch.
